pyproject.toml
[project]
name = "rag-ddd"
version = "0.1.0"
description = "Production-ready RAG with DDD architecture"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Your Team" }]

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "openai>=1.40.0",
  "qdrant-client>=1.9.0",
  "redis>=5.0.0",
  "pymongo>=4.7.0",
  "google-cloud-storage>=2.17.0",
  "tenacity>=9.0.0",
  "tiktoken>=0.7.0",
  "python-dotenv>=1.0.1",
  "structlog>=24.1.0",
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pypdf>=4.2.0",
  "opik>=1.0.0",
]

[project.scripts]
rag-cli = "rag_ddd.interface.cli.main:app"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
